add missing coverage
