+++ AtemSharp.Tests/Commands/Macro/MacroRecordCommandTests.cs
@@ -67,4 +67,10 @@ public class MacroRecordCommandTests : SerializedCommandTestBase<MacroRecordComm
     {
         TestPropertyNonMerging(Factory, property, firstValue, secondValue);
     }
+
+    [Test]
+    public void TestPropertyMerging_WithWrongType()
+    {
+        TestPropertyMerging_WithWrongType(Factory);
+    }
 }
